sdk38: Move proposal submission into test script
