+++ scripts/wasmd/init.sh
@@ -19,4 +19,3 @@ SCRIPT_DIR="$(realpath "$(dirname "$0")")"
 "$SCRIPT_DIR/deploy_erc20.js"
 "$SCRIPT_DIR/deploy_nameservice.js"
 "$SCRIPT_DIR/deploy_staking.js"
-"$SCRIPT_DIR/submit_proposal.js"
